--- src/test/java/ru/practicum/shareit/UserControllerTests.java
@@ -1,139 +1,0 @@
-package ru.practicum.shareit;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.jupiter.api.Test;
-import org.mockito.ArgumentMatchers;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
-import ru.practicum.shareit.user.UserController;
-import ru.practicum.shareit.user.dto.UserDto;
-import ru.practicum.shareit.user.interfaces.UserService;
-
-import java.nio.charset.StandardCharsets;
-import java.util.List;
-
-import static org.hamcrest.Matchers.*;
-import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.Mockito.when;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@WebMvcTest(controllers = UserController.class)
-public class UserControllerTests {
-
-    @MockBean
-    UserService userService;
-
-    @Autowired
-    ObjectMapper mapper;
-
-    @Autowired
-    private MockMvc mvc;
-
-    @Test
-    void addUserTest() throws Exception {
-
-        UserDto userDto = makeUserDto(1L, "User", "User@User.ru");
-
-        when(userService.addUser(ArgumentMatchers.any()))
-                .thenReturn(userDto);
-
-        mvc.perform(post("/users")
-                        .content(mapper.writeValueAsString(userDto))
-                        .characterEncoding(StandardCharsets.UTF_8)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.id", is(userDto.getId()), Long.class))
-                .andExpect(jsonPath("$.name", is(userDto.getName())))
-                .andExpect(jsonPath("$.email", is(userDto.getEmail())));
-    }
-
-    @Test
-    void updateUserTest() throws Exception {
-
-        long userId = 1L;
-        UserDto userDto = makeUserDto(userId, "User", "User@User.ru");
-
-        when(userService.updateUser(ArgumentMatchers.any()))
-                .thenReturn(userDto);
-
-        mvc.perform(patch("/users/" + userId)
-                        .content(mapper.writeValueAsString(userDto))
-                        .characterEncoding(StandardCharsets.UTF_8)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.id", is(userDto.getId()), Long.class))
-                .andExpect(jsonPath("$.name", is(userDto.getName())))
-                .andExpect(jsonPath("$.email", is(userDto.getEmail())));
-    }
-
-    @Test
-    void getAllUserTest() throws Exception {
-
-        long userId = 1L;
-        UserDto userDto1 = makeUserDto(userId, "User1", "User1@User.ru");
-        UserDto userDto2 = makeUserDto(userId + 1, "User2", "User2@User.ru");
-
-        List<UserDto> userDtoList = List.of(userDto1, userDto2);
-        when(userService.getAllUsers())
-                .thenReturn(userDtoList);
-
-        mvc.perform(get("/users/")
-                        .characterEncoding(StandardCharsets.UTF_8)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .accept(MediaType.APPLICATION_JSON))
-                .andExpect(jsonPath("$", hasSize(2)))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$[0].id", is(userDto1.getId()), Long.class))
-                .andExpect(jsonPath("$[0].name", is(userDto1.getName())))
-                .andExpect(jsonPath("$[0].email", is(userDto1.getEmail())))
-                .andExpect(jsonPath("$[1].id", is(userDto2.getId()), Long.class))
-                .andExpect(jsonPath("$[1].name", is(userDto2.getName())))
-                .andExpect(jsonPath("$[1].email", is(userDto2.getEmail())));
-    }
-
-    @Test
-    void getUserTest() throws Exception {
-
-        long userId = 1L;
-        UserDto userDto = makeUserDto(userId, "User", "User@User.ru");
-
-        when(userService.getUser(anyLong()))
-                .thenReturn(userDto);
-
-        mvc.perform(get("/users/" + userId)
-                        .characterEncoding(StandardCharsets.UTF_8)
-                        .accept(MediaType.ALL))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.id", is(userDto.getId()), Long.class))
-                .andExpect(jsonPath("$.name", is(userDto.getName())))
-                .andExpect(jsonPath("$.email", is(userDto.getEmail())));
-    }
-
-    @Test
-    void removeUserTest() throws Exception {
-
-        long userId = 1L;
-
-        mvc.perform(delete("/users/" + userId)
-                        .characterEncoding(StandardCharsets.UTF_8)
-                        .accept(MediaType.ALL))
-                .andExpect(status().isOk());
-    }
-
-    private UserDto makeUserDto(Long id, String name, String email) {
-
-        UserDto dto = new UserDto();
-        dto.setId(id);
-        dto.setEmail(email);
-        dto.setName(name);
-
-        return dto;
-    }
-}
