+++ src/test/java/ru/practicum/shareit/itemtests/dtojsontests/ItemDtoJsonTest.java
@@ -1,0 +1,43 @@
+package ru.practicum.shareit.itemtests.dtojsontests;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.json.JsonTest;
+import org.springframework.boot.test.json.JacksonTester;
+import org.springframework.boot.test.json.JsonContent;
+import ru.practicum.shareit.item.dto.ItemDto;
+
+import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
+
+@JsonTest
+public class ItemDtoJsonTest {
+
+    @Autowired
+    private JacksonTester<ItemDto> json;
+
+    @Test
+    void testItemDto() throws Exception {
+        ItemDto dto = new ItemDto();
+        dto.setId(1L);
+        dto.setName("Name");
+        dto.setDescription("Description");
+        dto.setAvailable(true);
+        dto.setOwner(2L);
+        dto.setRequestId(4L);
+
+        JsonContent<ItemDto> result = json.write(dto);
+
+        assertThat(result).extractingJsonPathNumberValue("$.id")
+                .isEqualTo(1);
+        assertThat(result).extractingJsonPathStringValue("$.name")
+                .isEqualTo("Name");
+        assertThat(result).extractingJsonPathStringValue("$.description")
+                .isEqualTo("Description");
+        assertThat(result).extractingJsonPathBooleanValue("$.available")
+                .isEqualTo(true);
+        assertThat(result).extractingJsonPathNumberValue("$.owner")
+                .isEqualTo(2);
+        assertThat(result).extractingJsonPathNumberValue("$.requestId")
+                .isEqualTo(4);
+    }
+}
